backend/requirements.txt
fastapi
uvicorn
sqlalchemy
firebase-admin
aiosqlite
python-dotenv